app/requirements.txt
streamlit==1.31.0
transformers==4.37.2
torch==2.1.2
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
nltk==3.8.1
python-dotenv==1.0.1
openai==1.12.0
